requirements.txt
python-telegram-bot==20.7
anthropic>=0.39.0
asyncpg>=0.29.0
gspread>=5.12.0
google-auth>=2.0.0
python-dotenv>=1.0.0
